Refactored MenuView and added header, hero section, menu breakdown, menu items views
